Revert "Extract Instabug to its own Module (#780)"
This reverts commit e477bed221c27be911bd1b8aee423406094f4ae3.
