+++ tests/index.spec.js
@@ -1,14 +1,13 @@
-import '../mocks/mockInstabugUtils';
+import './mocks/mockInstabugUtils';
 import React from 'react';
 import { NativeModules, Platform, processColor, findNodeHandle, Text } from 'react-native';
 import waitForExpect from 'wait-for-expect';
-import Instabug from '../../src/modules/Instabug';
-import Report from '../../src/models/Report';
-
-import IBGConstants from '../../src/utils/InstabugConstants';
-import IBGEventEmitter from '../../src/utils/IBGEventEmitter';
+import Instabug from '../src/';
+import Report from '../src/models/Report';
+import IBGConstants from '../src/utils/InstabugConstants';
+import IBGEventEmitter from '../src/utils/IBGEventEmitter';
 import { green } from 'ansi-colors';
-import InstabugUtils from '../../src/utils/InstabugUtils';
+import InstabugUtils from '../src/utils/InstabugUtils';
 
 const { Instabug: NativeInstabug } = NativeModules;
